Changed handling of non-existing JSON properties
Replaced Optionals with @Nullable annotation in API data models. As the TheTVDB.com Swagger documentation does not provide any information which fields are optional and may be omitted I added the @Nullable annotation to *all* non-collection fields. This aims to reduce unnecessary Optional-handling and hardens the application in case the remote API switches properties e.g. from mandatory to optional.