artwork-types: aligned implementation to API v.4.3.2 specifications
Integrated code from CompanyTypesAPI into the regular CompanyAPI and completely removed the distinct *types API. Removed the 'all' prefix from method names to align with the API Swagger documentation resource name. Renamed properties in CompanyType.java to match the actual values that are returned by the remote service within the JSON. Seems to be incorrect in API documentation. This also means that the special JSON handling of company types in CompanyDTO can be removed again.