Changed some field type in EpisodeDetails DTO
Switched type of 'studios' property from String to Studio DTO. Although this property is currently not declared by the API documentation it's unlikely that such studio information will just be a list of simple Strings. This is kind of an educated guess which type to be used but should be okay for now as we check the API documentation for updated information with each new version.