Fixed broken links to API documentation
With some recent OpenAPI update the 'tags' for each route have been changed to start with an upper-case
letter and to no use hyphens but blanks instead. Unfortunately these tags are part of the anchor URLs to
quickly jump to a specific resource in the online API documentation. Due to these changes most of the links
within the code did no longer work with regard to automatically scrolling down to the specific resource.
Changed links according to new OpenAPI resource 'tags'.