Added support for response content compression
Added request header to allow the remove service to compress its JSON
response payload in order to save bandwidth. If the API responds with a
proper "Content-Encoding" header, the actual response data will be
decompressed first before it is forwarded to the JSON deserializer.

This change was inspired by https://github.com/thetvdb/v4-api/issues/109